refactor: 최상위 테스트 클래스의 폴더 네이밍을 helper로 변경
