+++ src/test/java/doore/helper/IntegrationTest.java
@@ -1,4 +1,4 @@
-package doore.integration;
+package doore.helper;
 
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
