+++ src/test/java/doore/helper/RepositorySliceTest.java
@@ -1,4 +1,4 @@
-package doore.integration;
+package doore.helper;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
@@ -7,7 +7,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-public abstract class RepositoryIntegrationTest {
+public abstract class RepositorySliceTest {
 
     @PersistenceContext
     protected EntityManager em;
